+++ include/Pir_Tensors/Pir_Tensor.hpp
@@ -25,7 +25,7 @@ namespace  Piranha {
         Tensor Reshape(std::vector<int64_t> shape_) const;
         Tensor transpose() const;
         Tensor permute(const std::vector<int64_t>& order) const;
-        void all(int64_t v) const;
+        void all(float v) const;
         Tensor operator[](int64_t i) ;
